+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Add onTabChange handler in Accordion component.
+- Add onActivePanelChange handler in Accordion component.
 
 # [0.24.0] - 26-07-2019